Add Socket.IO for real-time location updates and serve index.html file
