+++ public/index.html
@@ -13,7 +13,7 @@
 <body>
   <div id="map"></div>
   <script src="/socket.io/socket.io.js"></script>
-  <script src="https://maps.googleapis.com/maps/api/js?key=YOUR_GOOGLE_MAPS_API_KEY"></script>
+  <script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyB5i-FYUFnAM4BRGChTXj2EaxsbuJIlALk"></script>
   <script>
     const socket = io();
 
@@ -44,5 +44,6 @@
 
     initMap();
   </script>
+  <script src="/main.js"></script>
 </body>
 </html>
