--- src/GPS-tracking/services/server.js
@@ -1,25 +1,0 @@
-// server.js
-const express = require('express');
-const http = require('http');
-const socketIo = require('socket.io');
-
-const app = express();
-const server = http.createServer(app);
-const io = socketIo(server);
-
-app.use(express.static('public'));
-
-io.on('connection', (socket) => {
-  console.log('New client connected');
-  
-  socket.on('locationUpdate', (location) => {
-    console.log('Location update received:', location);
-    io.emit('locationUpdate', location);
-  });
-
-  socket.on('disconnect', () => {
-    console.log('Client disconnected');
-  });
-});
-
-server.listen(3000, () => console.log('Listening on port 3000'));
